requirements.txt
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
opencv-python-headless==4.5.1.48
numpy==1.18.5
Werkzeug==1.0.1
requests-oauthlib==1.3.0